9043/refactor: extract Bag calculation logic into BagCalculatorService from PaymentCalculatorService
