+++ service-api/src/main/java/greencity/service/ubs/calculator/CertificateCalculatorService.java
@@ -1,4 +1,6 @@
 package greencity.service.ubs.calculator;
 
+//TODO add docs
+//TODO add tests
 public interface CertificateCalculatorService {
 }
